use stream logs strategy pattern
